Move pallet tests into a separate file
