+++ Cargo.toml
@@ -37,14 +37,13 @@ resolver = "2"
 edition = "2024"
 
 [workspace.lints.clippy]
-indexing-slicing = "deny"
 arithmetic-side-effects = "deny"
-type_complexity = "allow"
-unwrap-used = "deny"
+indexing-slicing = "deny"
 manual_inspect = "allow"
 result_large_err = "allow"
+type_complexity = "allow"
+unwrap-used = "deny"
 useless_conversion = "allow"     # until polkadot is patched
-result_large_err = "allow"
 
 [workspace.dependencies]
 node-subtensor-runtime = { default-features = false, path = "runtime" }
